Added more reporter methods

+++ entities/Exam.cs
@@ -4,7 +4,7 @@ namespace CoreSchool.entities
 {
     public class Exam:BaseSchoolObject
     {
-        public double grade { get; set; }
+        public float grade { get; set; }
         public Subject examSubject { get; set; }
         public Student testedStudent { get; set; }
